# Conflicts:
#	src/main/java/no/maoyi/app/user/entity/User.java

+++ src/main/java/no/maoyi/app/conversation/entity/Conversation.java
@@ -18,7 +18,7 @@ public class Conversation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    BigInteger id;
+    long id;
 
     @ManyToOne(cascade = CascadeType.ALL)
     Listing baseOrder;
